Outline border (https://github.com/niivue/ipyniivue/issues/220)

+++ src/ipyniivue/widget.py
@@ -220,7 +220,7 @@ class MeshLayer(BaseAnyWidget):
     use_negative_cmap = t.Bool(False).tag(sync=True)
     cal_min = t.Float(None, allow_none=True).tag(sync=True)
     cal_max = t.Float(None, allow_none=True).tag(sync=True)
-    outline_border = t.Int(0).tag(sync=True)
+    outline_border = t.Float(0).tag(sync=True)
 
     # other properties that aren't in init
     colormap_invert = t.Bool(False).tag(sync=True)
